requirements.txt
# AWS
boto3>=1.34.0

# PDF processing
pypdf>=4.0.0

# Anthropic Claude API
anthropic>=0.40.0

# LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0

# Embeddings
sentence-transformers>=3.0.0

# Vector store
faiss-cpu>=1.8.0

# HuggingFace
huggingface-hub>=0.24.0

# ML dependencies
numpy==1.26.4

# Environment variables
python-dotenv>=1.0.0

streamlit>=1.40.0